functions/requirements.txt
# Firebase Functions
firebase-functions>=0.4.0
firebase-admin>=6.5.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Google Cloud
google-cloud-firestore>=2.16.0
google-cloud-secret-manager>=2.20.0

# HTTP & API Clients
httpx>=0.28.0
spotipy>=2.24.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# String Matching
rapidfuzz>=3.10.0

# AI & MCP
anthropic>=0.39.0
claude-agent-sdk>=0.1.0
mcp>=1.0.0

# Required for agent execution
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
openai>=1.55.0
